refactor(canvas): move canvas types
