+++ src/renderer/src/features/chat/components/canvas/canvas-content.tsx
@@ -1,8 +1,8 @@
 import type { FC } from "react";
+import type { CanvasDocument } from "../../types/canvas";
 import { isMarkdownFile } from "../../utils/is-markdown-file";
 import { CodeEditor } from "./code-editor";
 import { MarkdownPreview } from "./markdown-preview";
-import type { CanvasDocument } from "./types";
 
 type CanvasContentProps = {
 	/**
