+++ src/renderer/src/features/chat/components/canvas/canvas-tabs.tsx
@@ -2,7 +2,7 @@ import { faTimes } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Box, Tab, Tabs, alpha, styled } from "@mui/material";
 import { type FC, useCallback, useEffect, useState } from "react";
-import type { CanvasDocument } from "./types";
+import type { CanvasDocument } from "../../types/canvas";
 
 type CanvasTabsProps = {
 	/**
