+++ src/renderer/src/features/chat/components/message-input.tsx
@@ -18,7 +18,7 @@ import { styled } from "@mui/material/styles";
 import { useMessageInput } from "@shared/hooks/use-message-input";
 import { useMemo, useRef, useState } from "react";
 import type { ChangeEvent, FC, FormEvent } from "react";
-import type { Message } from "../types";
+import type { Message } from "../types/message";
 import { AttachmentsPreview } from "./attachments-preview";
 import { ScrollToBottomButton } from "./scroll-to-bottom-button";
