+++ src/renderer/src/features/chat/components/message-item/index.tsx
@@ -2,6 +2,7 @@ import { Box } from "@mui/material";
 import { createLocalOperatorClient } from "@shared/api/local-operator";
 import { apiConfig } from "@shared/config";
 import { type FC, memo, useCallback, useMemo } from "react";
+import type { Message } from "../../types/message";
 import { ActionHighlight } from "./action-highlight";
 import { BackgroundBlock } from "./background-block";
 import { CodeBlock } from "./code-block";
@@ -18,7 +19,6 @@ import { MessageTimestamp } from "./message-timestamp";
 import { OutputBlock } from "./output-block";
 import { SecurityCheckHighlight } from "./security-check-highlight";
 import { StatusIndicator } from "./status-indicator";
-import type { Message } from "../../types";
 import { VideoAttachment } from "./video-attachment";
 
 /**
