+++ src/renderer/src/features/chat/components/message-item/message-paper.tsx
@@ -2,7 +2,7 @@ import { Box, Paper, useTheme } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import { useScrollToBottom } from "@shared/hooks/use-scroll-to-bottom";
 import { type FC, useEffect, useMemo, useRef } from "react";
-import type { Message } from "../../types";
+import type { Message } from "../../types/message";
 import { MessageControls } from "./message-controls";
 import { MessageTimestamp } from "./message-timestamp";
 import { StreamingMessage } from "./streaming-message";
