+++ src/renderer/src/features/chat/components/messages-view.tsx
@@ -4,7 +4,7 @@ import type {
 	JobStatus,
 } from "@shared/api/local-operator/types";
 import type { FC, RefObject } from "react";
-import type { Message } from "../types";
+import type { Message } from "../types/message";
 import { LoadingIndicator } from "./loading-indicator";
 import { MessageItem } from "./message-item";
