+++ src/renderer/src/features/chat/types/canvas.ts
