+++ src/renderer/src/features/chat/types/index.ts
@@ -1,0 +1,1 @@
+export type { Message } from "./message";
